Match grouping against q-matrix

Compares qmatrix output for a chosen value of K against a target_group vector. Returns the order of target_group groups, such that there is the best possible alignment against the qmatrix. For example, if the vector returned is c(2,3,1) then the second group in the target vector should be matched against the first group in the qmatrix, followed by the third group in the target vector against the second group in the qmatrix, followed by the first group in the target vector against the third group in the qmatrix.

get_group_order(project, K, target_group)

Arguments

project
a MALCOT project, as produced by the function malecot_project()

K
compare against qmatrix output for this value of K

target_group
the target group to be aligned against the qmatrix

Examples

# TODO