Simulate genetic data

Generates simulated data in the format accepted by MALECOT. Simulated data can be either bi-allelic or multi-allelic. In addition to returning the simulated data set, this function returns intermediate outputs used in data generation, including the true complexity of infection and true subpopulation membership of all samples. These additional outputs will not generally be present in real-world data sets, but can be useful in assessing the accuracy of the MCMC inference step by "truthing" estimated values (this is one of the major advantages of using simulated data).

sim_data(n = 100, L = 24, K = 3, data_format = "biallelic",
  alleles = 2, lambda = 1, COI_model = "poisson", COI_manual = NULL,
  COI_max = 20, COI_mean = 3, COI_dispersion = 1, e1 = 0, e2 = 0,
  prop_missing = 0)

Arguments

n
number of samples

L
number of loci

K
number of subpopulations. Samples are drawn with equal probability from each of these subpopulations

data_format
character string indicating whether data should be "biallelic" or "multiallelic". In the bi-allelic case data are formatted similar to The Real McCOIL (Chang et al. 2017). In the multi-allelic case data are in long format, allowing a different number of haplotypes per sample

alleles
vector of length L indicating the number of possible alleles at each locus, or if a single value is used then the same value applies for all loci. In the bi-allelic case this parameter is ignored as the number of alleles is fixed at 2

lambda
shape parameter(s) governing the prior on allele frequencies. This prior is a Beta distribution for the bi-allelic case or a Dirichlet distribution for the multi-allelic case. In the simplest case, lambda can be a single scalar value, in which case the same shape paremeter will be used for all loci and all alleles (i.e. a symmetric Beta or Dirichlet prior). In the more general case lambda can be a list of length L, containing vectors of length equal to the number of alleles at each locus, allowing different shape parameters to be defined for each allele individually

COI_model
the type of prior distribution on COI in each subpopulation. Can be uniform, poisson or nb (negative binomial). Note that the poisson and negative binomial distributions use a non-standard parametrization to account for the fact that COI=0 is not possible (see Details). If COI_manual is used then this prior is ignored.

COI_manual
manually define COI of all samples

COI_max
the maximum possible COI in any individual, irrespective of the type of prior on COI

COI_mean
the mean COI in a given subpopulation (applies to poisson and nb priors only). If vector valued then should be of length K and defines the mean separately for each subpopulation. If scalar valued then the same value applies to all subpopulations. Note that this mean value will only be correct when COI_max is sufficiently large (see Details)

COI_dispersion
the index of dispersion (i.e. the variance over the mean) of the prior on COI. Applies to nb prior only - in the Poisson case the dispersion is fully defined by the COI_mean. If vector valued then should be of length K and defines the dispersion separately for each subpopulation. If scalar valued then the same value applies to all subpopulations. Note that this dispersion will only be correct when COI_max is sufficiently large (see Details)

e1
first error term: the probabilty of a true homozygous locus being mis-called as a heterozygote. Errors are only allowed under the biallelic model

e2
second error term: the probabilty of a true heterozygous locus being mis-called as a homozygote. Errors are only allowed under the biallelic model

prop_missing
the proportion of missing data. Missing data is randomly distributed between samples and loci, and is coded as -1

Details

The sim_data() function returns a list containing multiple elements The main output is the element data, which is the final simulated data set, along with the number of samples (n) and the number of loci (L). If the parameters e1 or e2 are non-zero, or if prop_missing is greater than zero, then an uncorrupted version of the data is also returned in the element data_uncorrupted. This object represents what the true underlying data looked like before errors and missing data were introduced. Other "truthing" elements include the true subpopulation grouping, true COI, and true allele frequencies that were simulated to produce the final data. The accuracy of MCMC-based inference can be assessed by comparing posterior estimates against these "true" values used in simulation. Finally, the original call to the function is also stored in the element call for the sake of producing a complete record of all simulation parameters used.

The prior on COI can be uniform, Poisson, or negative binomial. In the uniform case there is an equal chance of any given sample having a COI between 1 and COI_max (inclusive). In the Poisson and negative binomial cases it is important to note that the distribution is over (COI-1), rather than over COI. This is because both Poisson and negative binomial distributions allow for 0 values, which cannot be the case here because observed samples must contain at least 1 genotype.

The full probability mass distribution for the Poisson case with COI_mean\(=\mu\) and COI_max\(=M\) can be written $$ Pr(COI=n) = z (\mu-1)^(n-1) exp(-(\mu-1)) / (n-1)! $$ where z is a normalising constant that ensures the distribution sums to unity: $$1/z = \sum_{i=1}^M (\mu-1)^(i-1) exp(-(\mu-1)) / (i-1)! $$

The mean of this distribution will generally be very close to \(\mu\), and the variance will be close to \(\mu-1\) (strictly it will approach these values as \(M\) tends to infinity).

The full probability mass distribution for the negative binomial case with COI_mean\(=\mu\), COI_dispersion\(=v/\mu\) and COI_max\(=M\) can be written $$ Pr(COI=n) = z \Gamma(n-1+N)/( \Gamma(N)(n-1)! ) p^N (1-p)^(n-1) $$ where \(N=(\mu-1)^2/(v-\mu+1)\), \(p=(\mu-1)/v\), and z is a normalising constant that ensures the distribution sums to unity: $$1/z = \sum_{i=1}^M \Gamma(i-1+N)/( \Gamma(N)(i-1)! ) p^N (1-p)^(i-1) $$

The mean of this distribution will generally be very close to \(\mu\), and the variance will be close to \(v\) (strictly it will approach these values as \(M\) tends to infinity).

References

Chang, H.H., Worby, C.J., Yeka, A., Nankabirwa, J., Kamya, M.R., Staedke, S.G., Dorsey, G., Murphy, M., Neafsey, D.E., Jeffreys, A.E. and Hubbart, C., 2017. THE REAL McCOIL: A method for the concurrent estimation of the complexity of infection and SNP allele frequency for malaria parasites. PLoS computational biology, 13(1), p.e1005348.

Examples

# TODO